refactor: extract SearchPage data logic into useSearchData hook (F5)
Move SearchPage's ~400 lines of fetch functions, state, refs, and effects
into a page-local hook (components/search/useSearchData.js) built on the
same request-id/AbortController patterns, and its shared pure helpers into
components/search/searchUtils.js. SearchPage becomes a thin render shell
(913 -> 447 lines) that calls useSearchData() and renders. The hook is now
independently testable and the page reads as pure presentation.

Behavior-preserving: the hook body is the component body moved verbatim
(same hook-call order, same effect deps), returning exactly the state and
handlers the JSX consumes; verified by SearchPage's existing 6-test suite,
the full frontend suite (83), lint, and build.
